move SetImplementationSourceExplorer to its own file
